Refactor : useClickOutside 공통 훅 생성 및 적용
